feat: inputs in SyncComputeRequest
Inputs are global to a SyncComputeRequest. Additionally, introduce a
`FHE_GET_INPUT_CIPHERTEXT` FHE operation that allows for a client to get
an input ciphertext (that is not part of FHE computation). Would be used
by the `verifyCiphertext` function in fhEVM-native.

Change ciphertext version to 0.

Add an integration test for gettting an input ciphertext.

Introduce ciphertext compression - all ciphertexts coming in and going
out of `SyncComputeRequest` will be compressed.