feat: make ciphertext inputs global per request
Instead of passing in ciphertexts as part of the SyncInput, pass them
globally for the whole request. Decompress them into a map that lives
for the duration of the request. Computations then refer to ciphertexts
in the map via handles, applying to inputs, compressed ciphertexts and
results.

Above also allows for use of results as inputs in subsequent
computations.