feat: use versioned safe ser/deser
Use the versioned safe ser/deser tfhe-rs functions. Versioning might
become part of safe ser/deser, but for now we better be explicit and not
rely on that.

`ProvenCompactCiphertextList` uses safe_(de)serialize() now as
versioning is not supported in thfe-rs yet. Will move to versioned once
available.

`CompactPkePublicParams` uses (de)serialize_with_mode() as of now as
versioning is not supported yet.

Also, remove explicit thread local and Arc handling for server key - it
is already an Arc internally, so it is cheap to clone.