feat: sign handle and ciphertext digest in copro
Sign the handle and a keccak256 digest of the ciphertext when returning
a GetCiphertextResponse in the coprocessor.

The EIP712 domain used is:

```
let get_ciphertext_eip712_domain = alloy::sol_types::eip712_domain! {
    name: "GetCiphertextResponse",
    version: "1",
};
```

The data structure we sign is:

```
alloy::sol! {
    struct GetCiphertextResponseSignatureData {
        uint256 handle;
        bytes ciphertext_digest;
    }
}
```

The signature is returned in the new `FetchedCiphertext.signature` field.

Resolves https://github.com/zama-ai/fhevm/issues/599