feat: first version of input proofs in gw-listener
This commit adds input poofs handling in the gw-listener. It supports
restarting from the last seen block in case of a crash or stop of the
service.

We assume that the ZKPoKManager can handle proof requests that were
already completed, meaning that redoing some requests won't have
negative effects. That allows us to only remember the last **seen**
block number. For more information on the ZKPoKManager, see:

```@dev This means a "late" response will not be reverted, just ignored```

https://github.com/zama-ai/gateway-l2/blob/main/contracts/ZKPoKManager.sol#L125

Error handling will be improved in future commits. More tests will be
added too.

Reorg handling is also a topic for a future commit.