fix: read sns_pk large object from tenants table in chunks
The sns_pk field corresponds to the `ServerKey` with `enable_noise_squashing` enabled.
Its size (~1.1 GB) exceeds the limits of the `BYTEA` type, so it is stored as
a PostgreSQL Large Object (LO).

Directly reading the LO using `lo_get` fails due to size constraints. To handle this,
the `read_in_chunks` method is used to stream the object in chunks.